Add extra warnings.
Add the -Wextra for all targets, but also add -Wconversion for
the main code.

There was a bug where readlink was being assigned to a size_t and
this option would have detected that problem. So adding these
options to avoid this in the future.

Fix the miscompares found with the new options.

Test: Compiles. Unit tests pass.

+++ Options.cpp
@@ -213,14 +213,14 @@ bool Options::HandleArg(const std::string& arg, const std::string& value, const 
 }
 
 static bool ReadFileToString(const std::string& file, std::string* contents) {
-  int fd = TEMP_FAILURE_RETRY(open(file.c_str(), O_RDONLY | O_CLOEXEC));
+  int fd = static_cast<int>(TEMP_FAILURE_RETRY(open(file.c_str(), O_RDONLY | O_CLOEXEC)));
   if (fd == -1) {
     return false;
   }
   char buf[4096];
   ssize_t bytes_read;
   while ((bytes_read = TEMP_FAILURE_RETRY(read(fd, &buf, sizeof(buf)))) > 0) {
-    contents->append(buf, bytes_read);
+    contents->append(buf, static_cast<size_t>(bytes_read));
   }
   close(fd);
   return true;
